renamed client-server utility functions
- canUseDom is now renamed to isClient
- isServer is now removed and used as !isClient
