--- project-base/storefront/helpers/canUseDom.ts
@@ -1,6 +1,0 @@
-export const canUseDom = (): boolean =>
-    !!(
-        typeof window !== 'undefined' &&
-        typeof window.document === 'object' &&
-        typeof window.document.createElement === 'function'
-    );
