+++ project-base/storefront/helpers/isClient.ts
@@ -1,0 +1,5 @@
+export const isClient = !!(
+    typeof window !== 'undefined' &&
+    typeof window.document === 'object' &&
+    typeof window.document.createElement === 'function'
+);
