+++ project-base/storefront/helpers/isRedirectedFromSsr.ts
@@ -1,7 +1,4 @@
 import { NextIncomingMessage } from 'next/dist/server/request-meta';
-import { canUseDom } from './canUseDom';
-
-export const isServer = (): boolean => !canUseDom();
 
 export const isRedirectedFromSsr = (requestHeader: NextIncomingMessage['headers']): boolean => {
     return requestHeader['x-nextjs-data'] !== '1';
