+++ project-base/storefront/pages/flags/[flagSlug].tsx
@@ -15,7 +15,7 @@ import { getMappedProductFilter } from 'helpers/filterOptions/getMappedProductFi
 import { useGtmFriendlyPageViewEvent } from 'gtm/helpers/eventFactories';
 import { getServerSidePropsWrapper } from 'helpers/serverSide/getServerSidePropsWrapper';
 import { initServerSideProps } from 'helpers/serverSide/initServerSideProps';
-import { isRedirectedFromSsr } from 'helpers/isServer';
+import { isRedirectedFromSsr } from 'helpers/isRedirectedFromSsr';
 import {
     FILTER_QUERY_PARAMETER_NAME,
     LOAD_MORE_QUERY_PARAMETER_NAME,
